+++ demisto_sdk/commands/init/tests/initiator_test.py
@@ -1,6 +1,5 @@
 import os
 from collections import OrderedDict, deque
-from datetime import datetime
 from os import listdir
 from pathlib import Path
 from typing import Callable
@@ -104,7 +103,6 @@ class TestCreateMetadata:
             'author': XSOAR_AUTHOR,
             'url': XSOAR_SUPPORT_URL,
             'email': '',
-            'created': datetime.utcnow().strftime(Initiator.DATE_FORMAT),
             'categories': [],
             'tags': [],
             'useCases': [],
@@ -137,7 +135,6 @@ class TestCreateMetadata:
             'author': XSOAR_AUTHOR,
             'url': XSOAR_SUPPORT_URL,
             'email': '',
-            'created': datetime.utcnow().strftime(Initiator.DATE_FORMAT),
             'categories': [],
             'tags': [],
             'useCases': [],
@@ -178,7 +175,6 @@ class TestCreateMetadata:
             'name': PACK_NAME,
             'support': PACK_SUPPORT_OPTIONS[1],
             'tags': ['Tag1', 'Tag2'],
-            'created': datetime.utcnow().strftime(Initiator.DATE_FORMAT),
             'url': PACK_URL,
             'useCases': [],
             'githubUser': []
@@ -218,7 +214,6 @@ class TestCreateMetadata:
             'name': PACK_NAME,
             'support': PACK_SUPPORT_OPTIONS[1],
             'tags': ['Tag1', 'Tag2'],
-            'created': datetime.utcnow().strftime(Initiator.DATE_FORMAT),
             'url': PACK_URL,
             'useCases': [],
             'githubUser': []
@@ -258,7 +253,6 @@ class TestCreateMetadata:
             'name': PACK_NAME,
             'support': PACK_SUPPORT_OPTIONS[3],
             'tags': ['Tag1', 'Tag2'],
-            'created': datetime.utcnow().strftime(Initiator.DATE_FORMAT),
             'url': MARKETPLACE_LIVE_DISCUSSIONS,
             'useCases': [],
             'githubUser': []